attendedsysupgrade: add package
this package is a first "PoC" for a simple command line sysupgrade.
The upgrade server [1] is requested for a new version, if available the
user is ask to request and download an image fitting the installed
architecture.

This is only a quick draft, not meant to be used yet.

Missing:

- Support to read locally installed packages
- Verify downloaded images by signature and checksum
- Obey uci options

Signed-off-by: Paul Spooren <paul@spooren.de>