prometheus-alertmanager: new package
From the [official GitHub
page](https://github.com/prometheus/alertmanager/):
The Alertmanager handles alerts sent by client applications such as the
Prometheus server. It takes care of deduplicating, grouping, and routing
them to the correct receiver integrations such as email, PagerDuty, or
OpsGenie.  It also takes care of silencing and inhibition of alerts.

---

As prometheus-alertmanager is written in Go it requires the golang
package to be installed. Currently is golang still a pending pull
request[1].

Signed-off-by: Paul Spooren <spooren@informatik.uni-leipzig.de>